requirements.txt
# requirements.txt

# For speech recognition
SpeechRecognition

# For text-to-speech
pyttsx3

# For checking internet connection
requests

# For integrating with the Gemini API
google-generativeai

# For offline model inference
numpy
scikit-learn

# For semantic search for offline responses
sentence-transformers